backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
dnspython==2.7.0
python-whois==0.9.5
reportlab==4.2.5
pydantic==2.10.6
pydantic-settings==2.7.1
openai==1.68.2
python-multipart==0.0.20
gunicorn==23.0.0
